simulator/requirements.txt
confluent-kafka==2.5.0
websocket-client==1.8.0
requests==2.32.3